refactor: modules under 500 lines + fully config-driven large-file threshold
Two coupled maintainability changes that thread through baseline/create.ts, so
they land together.

Module extractions (Rule 2 — one cohesive unit per file, each now <500):
- baseline/policy.ts → baseline/classify.ts (ClassifyContext/ClassifyResult +
  classify/classifyAll)
- baseline/ref-baseline.ts → baseline/remote-ref.ts (withRemoteRefWorktree — the
  cross-repo sibling of withRefWorktree; clone/fetch, not a worktree, so it sits
  outside the Rule 11 git-worktree confinement)
- baseline/create.ts → baseline/tool-versions.ts (buildToolsMap + version cache)
- loop/stop-gate.ts → loop/repair-message.ts (blockingPairs / findingBreakdown /
  buildRepairMessage)
- discovery/commands.ts → command-types.ts (types) + advisor.ts (pure doctor
  probes + config planners) + command-defs.ts (COMMANDS data); commands.ts is now
  a thin facade re-exporting the registry + helpers, so every importer of
  ./commands is unchanged. The Rule 16 arch-check reads command-defs.ts for the
  id/alias parity set.

Config-driven large-file threshold (one concept, one code path — CLAUDE.md
Rule 2.30). Previously the threshold lived hardcoded at TWO independent sites:
the `filesOver500Lines` count + `largestFiles` list in the metrics gather, and a
just-added producer-level filter. That is exactly the two-code-path divergence
the rule warns about — a policy override would have moved the guardrail finding
but left the score + report pinned at 500.

Now the threshold has ONE application point: `gatherGenericMetrics` derives both
the count and `largestFiles` against it and records the resolved value on
`HealthMetrics.largeFileThreshold`. It is resolved once from
`policy.json:largeFileThreshold` at the analyzer boundary (default 500, validated
positive-finite) and threaded in. Every consumer — the baseline `large-file`
producer, the Quality + Maintainability scores, and the "over N lines" prose —
reads those recorded fields, so they cannot disagree. The producer's throwaway
threshold param + ProducerContext field are removed; the producer is a pure 1:1
projection of the already-filtered list. Documented in docs/configuration/policy.md.

With the five files back under 500, their net-new large-file findings disappear,
so the five large-file allowlist entries are dropped (49 remain). New gather-level
test proves the threshold drives count + list + record together across values;
producer test asserts the 1:1 projection contract. Full suite 3320 green;
typecheck, arch, lint, prettier clean.
