+++ src/baseline/check.ts
@@ -56,8 +56,10 @@ import { gitAwareMatch } from './git-aware-match';
 import type { LocatedIdentity } from './git-aware-match';
 import { resolveBaselineMode } from './modes';
 import type { ResolvedMode } from './modes';
-import { classify, resolvePolicy, loadPolicyFromCwd } from './policy';
-import type { BrownfieldPolicy, ClassifyContext, ClassifyResult, BaselineSection } from './policy';
+import { resolvePolicy, loadPolicyFromCwd } from './policy';
+import { classify } from './classify';
+import type { BrownfieldPolicy, BaselineSection } from './policy';
+import type { ClassifyContext, ClassifyResult } from './classify';
 import { hydrateAnchorFromBranch, loadAnchorFromBranch } from './anchor';
 import { gatherFromRef } from './ref-baseline';
 import { type GatherScope, FULL_SCOPE, scopeForPolicy } from './gather-scope';
