+++ test/quality-scoring.test.ts
@@ -26,6 +26,7 @@ function clean(): QualityScoreInput {
     staleFiles: 0,
     mixedLanguages: false,
     filesOver500Lines: 0,
+    largeFileThreshold: 500,
     largestFileLines: 0,
     anyTypeCount: 0,
     typeErrors: null,
@@ -50,6 +51,7 @@ describe('scoreQualityFromInput — baseline', () => {
       ...clean(),
       lintErrors: 5000,
       filesOver500Lines: 100,
+      largeFileThreshold: 500,
       largestFileLines: 20000,
       consoleLogCount: 500,
       anyTypeCount: 5000,
